Use int64_t for gcd()

+++ src/rf.c
@@ -353,9 +353,9 @@ int rf_mixer_process(struct rf_mixer_t *s, int16_t *out, const int16_t *in, unsi
 	return(0);
 }
 
-int rf_gcd(int a, int b)
+int64_t rf_gcd(int64_t a, int64_t b)
 {
-	int c;
+	int64_t c;
 	
 	while((c = a % b))
 	{
